requirements.txt
asgiref==3.3.1
Django==3.1.7
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
